test(#1274): re-home orphaned Matérn n-free re-key gate into gam-models + pin honest contract
The #1274 fix (6a5a2e1) was reverted ~2h later by feb0eb5 (dropped Matérn from
supports_nfree_penalty_rekey) without reopening the issue. Its regression test
was orphaned by #1601 — commented out in gam_terms::smooth::tests and needing
the gam-models-private FrozenTermCollectionIncrementalRealizer it could not see
— so it compiled nowhere AND asserted the opposite of shipped code.

Re-homed into crates/gam-models/.../drivers/matern_nfree_rekey_topology_tests.rs
(same #1601 re-home pattern as the #901 κ tests). Tests 1/3/4 pin the n-free
operator-triplet penalty re-key as byte/topology-identical to the slow path
across ψ (this half of the fix is real and survives). The admission test now
honestly pins the POST-REVERT production contract: Matérn is NOT admitted to the
design-realization skip (slow-but-right), with a message pointing at the mgcv
quality bar a future re-enable must re-verify. No XFAIL, no orphaning.
